+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - Add support for `icrc21_canister_call_consent_message` to `@dfinity/ledger-icp` and `@dfinity/ledger-icrc`.
+- Add support for `"regtest"` in `BitcoinNetwork`.
 
 # 2024.09.02-0830Z