+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Add support for `icrc21_canister_call_consent_message` to `@dfinity/ledger-icp` and `@dfinity/ledger-icrc`.
 - Add support for `"regtest"` in `BitcoinNetwork`.
 
+## Build
+
+- Incorporate Agent-js patch `v2.1.2`.
+
 # 2024.09.02-0830Z
 
 ## Overview