+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-# Next
+# 2024.09.30-1100Z
+
+## Overview
+
+The current status of the libraries at the time of the release is as follows:
+
+| Library                  | Version | Status              |
+| ------------------------ | ------- | ------------------- |
+| `@dfinity/ckbtc`         | v3.1.0  | Enhanced 🔧️        |
+| `@dfinity/cketh`         | v3.3.1  | Maintained ⚙️       |
+| `@dfinity/cmc`           | v3.2.1  | Maintained ⚙️       |
+| `@dfinity/ic-management` | v5.2.1  | Maintained ⚙️       |
+| `@dfinity/ledger-icp`    | v2.6.0  | Enhanced 🔧️        |
+| `@dfinity/ledger-icrc`   | v2.6.0  | Enhanced 🔧️        |
+| `@dfinity/nns`           | v7.0.0  | Breaking Changes ⚠️ |
+| `@dfinity/nns-proto`     | v2.0.1  | Unchanged️          |
+| `@dfinity/sns`           | v3.2.1  | Maintained ⚙️       |
+| `@dfinity/utils`         | v2.5.0  | Maintained ⚙️       |
 
 ## Breaking changes